+++ RichString.c
@@ -21,7 +21,7 @@ in the source distribution for its full text.
 
 #define charBytes(n) (sizeof(CharType) * (n))
 
-static void RichString_extendLen(RichString* this, int len) {
+static void RichString_extendLen(RichString* this, size_t len) {
    if (this->chptr == this->chstr) {
       // String is in internal buffer
       if (len > RICHSTRING_MAXLEN) {
@@ -49,7 +49,7 @@ static void RichString_extendLen(RichString* this, int len) {
    this->chlen = len;
 }
 
-static void RichString_setLen(RichString* this, int len) {
+static void RichString_setLen(RichString* this, size_t len) {
    if (len <= RICHSTRING_MAXLEN && this->chlen <= RICHSTRING_MAXLEN) {
       RichString_setChar(this, len, 0);
       this->chlen = len;
@@ -58,7 +58,7 @@ static void RichString_setLen(RichString* this, int len) {
    }
 }
 
-void RichString_rewind(RichString* this, int count) {
+void RichString_rewind(RichString* this, size_t count) {
    RichString_setLen(this, this->chlen > count ? this->chlen - count : 0);
 }
 
@@ -95,33 +95,33 @@ static size_t mbstowcs_nonfatal(wchar_t* restrict dest, const char* restrict src
    return written;
 }
 
-static inline int RichString_writeFromWide(RichString* this, int attrs, const char* data_c, int from, int len) {
+static inline size_t RichString_writeFromWide(RichString* this, int attrs, const char* data_c, size_t from, size_t len) {
    wchar_t data[len];
    len = mbstowcs_nonfatal(data, data_c, len);
    if (len <= 0)
       return 0;
 
-   int newLen = from + len;
+   size_t newLen = from + len;
    RichString_setLen(this, newLen);
-   for (int i = from, j = 0; i < newLen; i++, j++) {
+   for (size_t i = from, j = 0; i < newLen; i++, j++) {
       this->chptr[i] = (CharType) { .attr = attrs & 0xffffff, .chars = { (iswprint(data[j]) ? data[j] : L'\xFFFD') } };
    }
 
    return len;
 }
 
-int RichString_appendnWideColumns(RichString* this, int attrs, const char* data_c, int len, int* columns) {
+size_t RichString_appendnWideColumns(RichString* this, int attrs, const char* data_c, size_t len, int* columns) {
    wchar_t data[len];
    len = mbstowcs_nonfatal(data, data_c, len);
    if (len <= 0)
       return 0;
 
-   int from = this->chlen;
-   int newLen = from + len;
+   size_t from = this->chlen;
+   size_t newLen = from + len;
    RichString_setLen(this, newLen);
    int columnsWritten = 0;
-   int pos = from;
-   for (int j = 0; j < len; j++) {
+   size_t pos = from;
+   for (size_t j = 0; j < len; j++) {
       wchar_t c = iswprint(data[j]) ? data[j] : L'\xFFFD';
       int cwidth = wcwidth(c);
       if (cwidth > *columns)
@@ -140,10 +140,10 @@ int RichString_appendnWideColumns(RichString* this, int attrs, const char* data_
    return pos - from;
 }
 
-static inline int RichString_writeFromAscii(RichString* this, int attrs, const char* data, int from, int len) {
-   int newLen = from + len;
+static inline size_t RichString_writeFromAscii(RichString* this, int attrs, const char* data, size_t from, size_t len) {
+   size_t newLen = from + len;
    RichString_setLen(this, newLen);
-   for (int i = from, j = 0; i < newLen; i++, j++) {
+   for (size_t i = from, j = 0; i < newLen; i++, j++) {
       assert((unsigned char)data[j] <= SCHAR_MAX);
       this->chptr[i] = (CharType) { .attr = attrs & 0xffffff, .chars = { (isprint((unsigned char)data[j]) ? data[j] : L'\xFFFD') } };
    }
@@ -151,39 +151,39 @@ static inline int RichString_writeFromAscii(RichString* this, int attrs, const c
    return len;
 }
 
-inline void RichString_setAttrn(RichString* this, int attrs, int start, int charcount) {
-   int end = CLAMP(start + charcount, 0, this->chlen);
-   for (int i = start; i < end; i++) {
+inline void RichString_setAttrn(RichString* this, int attrs, size_t start, size_t charcount) {
+   size_t end = CLAMP(start + charcount, 0, this->chlen);
+   for (size_t i = start; i < end; i++) {
       this->chptr[i].attr = attrs;
    }
 }
 
-void RichString_appendChr(RichString* this, int attrs, char c, int count) {
-   int from = this->chlen;
-   int newLen = from + count;
+void RichString_appendChr(RichString* this, int attrs, char c, size_t count) {
+   size_t from = this->chlen;
+   size_t newLen = from + count;
    RichString_setLen(this, newLen);
-   for (int i = from; i < newLen; i++) {
+   for (size_t i = from; i < newLen; i++) {
       this->chptr[i] = (CharType) { .attr = attrs, .chars = { c, 0 } };
    }
 }
 
-int RichString_findChar(const RichString* this, char c, int start) {
+size_t RichString_findChar(const RichString* this, char c, size_t start) {
    const wchar_t wc = btowc(c);
    const cchar_t* ch = this->chptr + start;
-   for (int i = start; i < this->chlen; i++) {
+   for (size_t i = start; i < this->chlen; i++) {
       if (ch->chars[0] == wc)
          return i;
       ch++;
    }
-   return -1;
+   return (size_t)-1;
 }
 
 #else /* HAVE_LIBNCURSESW */
 
-static inline int RichString_writeFromWide(RichString* this, int attrs, const char* data_c, int from, int len) {
-   int newLen = from + len;
+static inline size_t RichString_writeFromWide(RichString* this, int attrs, const char* data_c, size_t from, size_t len) {
+   size_t newLen = from + len;
    RichString_setLen(this, newLen);
-   for (int i = from, j = 0; i < newLen; i++, j++) {
+   for (size_t i = from, j = 0; i < newLen; i++, j++) {
       this->chptr[i] = (((unsigned char)data_c[j]) >= 32 ? ((unsigned char)data_c[j]) : '?') | attrs;
    }
    this->chptr[newLen] = 0;
@@ -191,40 +191,40 @@ static inline int RichString_writeFromWide(RichString* this, int attrs, const ch
    return len;
 }
 
-int RichString_appendnWideColumns(RichString* this, int attrs, const char* data_c, int len, int* columns) {
-   int written = RichString_writeFromWide(this, attrs, data_c, this->chlen, MINIMUM(len, *columns));
-   *columns = written;
+size_t RichString_appendnWideColumns(RichString* this, int attrs, const char* data_c, size_t len, int* columns) {
+   size_t written = RichString_writeFromWide(this, attrs, data_c, this->chlen, MINIMUM(len, (unsigned int)*columns));
+   *columns = (int)written;
    return written;
 }
 
-static inline int RichString_writeFromAscii(RichString* this, int attrs, const char* data_c, int from, int len) {
+static inline size_t RichString_writeFromAscii(RichString* this, int attrs, const char* data_c, size_t from, size_t len) {
    return RichString_writeFromWide(this, attrs, data_c, from, len);
 }
 
-void RichString_setAttrn(RichString* this, int attrs, int start, int charcount) {
-   int end = CLAMP(start + charcount, 0, this->chlen);
-   for (int i = start; i < end; i++) {
+void RichString_setAttrn(RichString* this, int attrs, size_t start, size_t charcount) {
+   size_t end = CLAMP(start + charcount, 0, this->chlen);
+   for (size_t i = start; i < end; i++) {
       this->chptr[i] = (this->chptr[i] & 0xff) | attrs;
    }
 }
 
-void RichString_appendChr(RichString* this, int attrs, char c, int count) {
-   int from = this->chlen;
-   int newLen = from + count;
+void RichString_appendChr(RichString* this, int attrs, char c, size_t count) {
+   size_t from = this->chlen;
+   size_t newLen = from + count;
    RichString_setLen(this, newLen);
-   for (int i = from; i < newLen; i++) {
+   for (size_t i = from; i < newLen; i++) {
       this->chptr[i] = c | attrs;
    }
 }
 
-int RichString_findChar(const RichString* this, char c, int start) {
+size_t RichString_findChar(const RichString* this, char c, size_t start) {
    const chtype* ch = this->chptr + start;
-   for (int i = start; i < this->chlen; i++) {
+   for (size_t i = start; i < this->chlen; i++) {
       if ((*ch & 0xff) == (chtype) c)
          return i;
       ch++;
    }
-   return -1;
+   return (size_t)-1;
 }
 
 #endif /* HAVE_LIBNCURSESW */
@@ -240,26 +240,26 @@ void RichString_setAttr(RichString* this, int attrs) {
    RichString_setAttrn(this, attrs, 0, this->chlen);
 }
 
-int RichString_appendWide(RichString* this, int attrs, const char* data) {
+size_t RichString_appendWide(RichString* this, int attrs, const char* data) {
    return RichString_writeFromWide(this, attrs, data, this->chlen, strlen(data));
 }
 
-int RichString_appendnWide(RichString* this, int attrs, const char* data, int len) {
+size_t RichString_appendnWide(RichString* this, int attrs, const char* data, size_t len) {
    return RichString_writeFromWide(this, attrs, data, this->chlen, len);
 }
 
-int RichString_writeWide(RichString* this, int attrs, const char* data) {
+size_t RichString_writeWide(RichString* this, int attrs, const char* data) {
    return RichString_writeFromWide(this, attrs, data, 0, strlen(data));
 }
 
-int RichString_appendAscii(RichString* this, int attrs, const char* data) {
+size_t RichString_appendAscii(RichString* this, int attrs, const char* data) {
    return RichString_writeFromAscii(this, attrs, data, this->chlen, strlen(data));
 }
 
-int RichString_appendnAscii(RichString* this, int attrs, const char* data, int len) {
+size_t RichString_appendnAscii(RichString* this, int attrs, const char* data, size_t len) {
    return RichString_writeFromAscii(this, attrs, data, this->chlen, len);
 }
 
-int RichString_writeAscii(RichString* this, int attrs, const char* data) {
+size_t RichString_writeAscii(RichString* this, int attrs, const char* data) {
    return RichString_writeFromAscii(this, attrs, data, 0, strlen(data));
 }
