+++ RichString.h
@@ -7,6 +7,8 @@ Released under the GNU GPLv2+, see the COPYING file
 in the source distribution for its full text.
 */
 
+#include <stddef.h>
+
 #include "Macros.h"
 #include "ProvideCurses.h"
 
@@ -40,7 +42,7 @@ in the source distribution for its full text.
 #define RICHSTRING_MAXLEN 350
 
 typedef struct RichString_ {
-   int chlen;
+   size_t chlen;
    CharType* chptr;
    CharType chstr[RICHSTRING_MAXLEN + 1];
    int highlightAttr;
@@ -48,33 +50,33 @@ typedef struct RichString_ {
 
 void RichString_delete(RichString* this);
 
-void RichString_rewind(RichString* this, int count);
+void RichString_rewind(RichString* this, size_t count);
 
-void RichString_setAttrn(RichString* this, int attrs, int start, int charcount);
+void RichString_setAttrn(RichString* this, int attrs, size_t start, size_t charcount);
 
-int RichString_findChar(const RichString* this, char c, int start);
+size_t RichString_findChar(const RichString* this, char c, size_t start);
 
 void RichString_setAttr(RichString* this, int attrs);
 
-void RichString_appendChr(RichString* this, int attrs, char c, int count);
+void RichString_appendChr(RichString* this, int attrs, char c, size_t count);
 
 /* All appending and writing functions return the number of written characters (not columns). */
 
-int RichString_appendWide(RichString* this, int attrs, const char* data);
+size_t RichString_appendWide(RichString* this, int attrs, const char* data);
 
 ATTR_ACCESS3_R(3, 4)
-int RichString_appendnWide(RichString* this, int attrs, const char* data, int len);
+size_t RichString_appendnWide(RichString* this, int attrs, const char* data, size_t len);
 
 /* columns takes the maximum number of columns to write and contains on return the number of columns written. */
-int RichString_appendnWideColumns(RichString* this, int attrs, const char* data, int len, int* columns);
+size_t RichString_appendnWideColumns(RichString* this, int attrs, const char* data, size_t len, int* columns);
 
-int RichString_writeWide(RichString* this, int attrs, const char* data);
+size_t RichString_writeWide(RichString* this, int attrs, const char* data);
 
-int RichString_appendAscii(RichString* this, int attrs, const char* data);
+size_t RichString_appendAscii(RichString* this, int attrs, const char* data);
 
 ATTR_ACCESS3_R(3, 4)
-int RichString_appendnAscii(RichString* this, int attrs, const char* data, int len);
+size_t RichString_appendnAscii(RichString* this, int attrs, const char* data, size_t len);
 
-int RichString_writeAscii(RichString* this, int attrs, const char* data);
+size_t RichString_writeAscii(RichString* this, int attrs, const char* data);
 
 #endif
